+++ cmd/auth_service/internal/app/users.go
@@ -1,4 +1,4 @@
-package api
+package app
 
 import (
 	"github.com/gin-gonic/gin"
